requirements.txt
opencv-python>=4.5.0
mediapipe>=0.8.0
numpy>=1.19.0
spotipy>=2.19.0
python-vlc>=3.0.0 